+++ scripts/resources/ProjectileResource.gd
@@ -1,0 +1,7 @@
+extends Resource
+class_name ProjectileResource
+
+@export_category("Statistics")
+@export var life_time:= 5.0
+@export var distance := 600
+@export var gravity := 1.0