+++ info/ChangeLog.md
@@ -2,6 +2,11 @@
 
 Where changes are placed
 
+## 0.1.1
+
+* Added projectile resource
+* player can now shoot different projectiles
+
 ## 0.1
 
 * Added actor resource